Delete latest model from metadata.

+++ knuckle/storage/local.py
@@ -2,6 +2,7 @@ import asyncio
 import concurrent.futures
 import io
 import logging
+import operator
 import pathlib
 import shutil
 import tinydb
@@ -9,11 +10,13 @@ import typing
 
 import knuckle.errors
 import knuckle.logging
-import knuckle.storage.meta
 
 from knuckle import asynclib
 from knuckle import model
-from knuckle.storage.meta import query_by_name_and_tag, query_by_id
+from knuckle.storage import metadata
+from knuckle.storage.metadata import (query_by_name,
+                                      query_by_name_and_tag,
+                                      query_by_id)
 
 
 class FileSystem:
@@ -26,7 +29,7 @@ class FileSystem:
     @classmethod
     def new(cls,
             path: pathlib.Path,
-            meta: knuckle.storage.meta.DB,
+            meta: metadata.DB,
             loader: model.Loader,
             logger: logging.Logger=knuckle.logging.internal_logger):
 
@@ -38,12 +41,15 @@ class FileSystem:
         self.loader = loader
         self.models_path = path.joinpath("models")
 
+        self.on_delete = []
+        self.on_save = []
+
         self.models_path.mkdir(parents=True, exist_ok=True)
         self.executor = concurrent.futures.ThreadPoolExecutor()
 
         return self
 
-    def _new_model(self, document: typing.Dict) -> model.Model:
+    def build_model_from_document(self, document: typing.Dict) -> model.Model:
         path = self.models_path.joinpath(document["id"])
         return model.Model(path=path, loader=self.loader, **document)
 
@@ -63,6 +69,26 @@ class FileSystem:
             m = model.Model(path=path, loader=self.loader, **document)
             yield m
 
+    async def save_to_meta(self, m: model.Model) -> None:
+        async with self.meta.write_locked() as meta:
+            if await meta.get(query_by_name_and_tag(m.name, m.tag)):
+                self.logger.debug("Model %s already exists", m)
+
+                raise knuckle.errors.DuplicateError(m.name, m.tag)
+
+            # Insert the model metadata, and update the latest model link.
+            await meta.insert(m.to_dict())
+
+            # Since the saving is happening right now, the latest model
+            # will obviously be the current one.
+            latest = m.copy()
+
+            latest.tag = model.Tag.Latest.value
+            latest.id = m.id
+
+            latest_query = query_by_name_and_tag(latest.name, latest.tag)
+            await meta.upsert(latest.to_dict(), latest_query)
+
     async def save(self, name: str, tag: str,
                    stream: io.IOBase) -> model.Model:
         """Save the model into the local storage.
@@ -70,13 +96,6 @@ class FileSystem:
         Extracts the TAR archive into the data directory.
         """
         m = model.Model.new(name, tag, self.models_path, self.loader)
-
-        # Since the saving is happening right now, the latest model
-        # will obviously be the current one.
-        latest = m.copy()
-
-        latest.tag = model.Tag.Latest.value
-        latest.id = m.id
 
         try:
             task = asynclib.extract_tar(fileobj=stream, dest=m.path)
@@ -88,16 +107,7 @@ class FileSystem:
             task = asyncio.coroutine(m.load)()
             m = await self.await_in_thread(task)
 
-            async with self.meta.write_locked() as meta:
-                if await meta.get(query_by_name_and_tag(name, tag)):
-                    self.logger.debug("Model %s already exists", m)
-                    raise knuckle.errors.DuplicateError(name, tag)
-
-                # Insert the model metadata, and update the latest model link.
-                await meta.insert(m.to_dict())
-
-                latest_query = query_by_name_and_tag(latest.name, latest.tag)
-                await meta.upsert(latest.to_dict(), latest_query)
+            await self.save_to_meta(m)
 
             # Model successfully loaded, so now it can be moved to the original
             # data root directory.
@@ -115,31 +125,50 @@ class FileSystem:
             await self.await_in_thread(task)
             raise e
 
+    async def delete_from_meta(self, name: str, tag: str) -> model.Model:
+        # Model found, remove metadata from the database.
+        async with self.meta.write_locked() as meta:
+            m = await self.load_from_meta(name, tag)
+
+            await meta.remove(query_by_id(m.id))
+
+            # Remove the "latest" model link.
+            query = query_by_name_and_tag(m.name, model.Tag.Latest.value)
+            await meta.remove(query)
+
+            # Retrieve a new "latest" model.
+            key = operator.itemgetter("created_at")
+            document = await meta.latest(query_by_name(m.name), key)
+
+            latest = self.build_model_from_document(document)
+            latest.tag = model.Tag.Latest.value
+
+            await meta.insert(latest.to_dict())
+            return m
+
     async def delete(self, name: str, tag: str) -> None:
         """Remove model with the given name and tag."""
-        try:
-            # Model found, remove metadata from the database.
-            m = await self._load(name, tag)
-            await self.meta.remove(query_by_id(m.id))
+        if tag == model.Tag.Latest.value:
+            raise model.NotFoundError(name, tag)
 
-            # Remove the model data from the storage.
+        try:
+            # Remove the model from the metadata database.
+            m = await self.delete_from_meta(name, tag)
+
+            # Remove the model data from the file system.
             await self.await_in_thread(asynclib.remove_dir(m.path))
 
             self.logger.info("Removed model %s:%s", name, tag)
         except FileNotFoundError:
             raise knuckle.errors.NotFoundError(name, tag)
 
-    async def _load(self, name: str, tag: str):
+    async def load_from_meta(self, name: str, tag: str):
         document = await self.meta.get(query_by_name_and_tag(name, tag))
         if not document:
             raise knuckle.errors.NotFoundError(name, tag)
-        return self._new_model(document)
+        return self.build_model_from_document(document)
 
     async def load(self, name: str, tag: str) -> model.Model:
-        """Load model with the given name and tag.
-        
-        When the 'latest' tag is specified, the most recent model from the
-        group specified by name will be returned.
-        """
-        m = await self._load(name, tag)
+        """Load model with the given name and tag."""
+        m = await self.load_from_meta(name, tag)
         return await self.await_in_thread(asyncio.coroutine(m.load)())
